Add x64::Compiler
This allows for compiling WASM functions at runtime. This will serve as the basis for the API for platform specific compilers.

We're using the standard unix calling convensions and in general are emitting the simplest possible assembly. There are plenty of optimizations to perform.

Arm64 support will follow. We will only support these two architectures (linux or mac).